adding task completed check before attempting to kill allocation

+++ master/internal/command/notebook_manager.go
@@ -1,9 +1,11 @@
 package command
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/determined-ai/determined/master/internal/api"
+	"github.com/determined-ai/determined/master/internal/db"
 	"github.com/determined-ai/determined/master/internal/task"
 	"github.com/determined-ai/determined/master/pkg/model"
 	"github.com/determined-ai/determined/master/pkg/protoutils"
@@ -65,15 +67,25 @@ func (cs *CommandService) KillNotebook(req *apiv1.KillNotebookRequest) (*apiv1.K
 	cs.mu.Lock()
 	defer cs.mu.Unlock()
 
-	c, err := cs.getNTSC(model.TaskID(req.NotebookId), model.TaskTypeNotebook)
+	tID := model.TaskID(req.NotebookId)
+
+	c, err := cs.getNTSC(tID, model.TaskTypeNotebook)
 	if err != nil {
 		return nil, err
 	}
 
-	err = task.DefaultService.Signal(c.allocationID, task.KillAllocation, "user requested kill")
+	completed, err := db.TaskCompleted(context.TODO(), tID)
 	if err != nil {
-		return nil, fmt.Errorf("failed to kill allocation: %w", err)
+		return nil, err
 	}
+
+	if !completed {
+		err = task.DefaultService.Signal(c.allocationID, task.KillAllocation, "user requested kill")
+		if err != nil {
+			return nil, fmt.Errorf("failed to kill allocation: %w", err)
+		}
+	}
+
 	return &apiv1.KillNotebookResponse{Notebook: c.toNotebook()}, nil
 }
